Debug + should_trace fixes
Debug for pooled strings no longer waits for the collector. This is
important if Debug is called from a Drop impl, for example.

The should trace bit was being obliterated as part of allocation reuse.
Rather than fixing it, I realized that we didn't need to store the bit
in the first place since we always have the type.